+++ alinkjs/packages/assetlink-default-plugins/webpack.config.js
@@ -6,6 +6,7 @@ const yaml = require('js-yaml');
 const path = require('path');
 const zlib = require('zlib');
 const CopyPlugin = require("copy-webpack-plugin");
+const { GenerateDefaultPluginConfigYmlFilesPlugin } = require('assetlink-plugin-dev-support');
 
 
 const DEV_PROXY_TARGET = process.env.ASSET_LINK_DEV_PROXY_TARGET || 'http://127.0.0.1';
@@ -178,42 +179,6 @@ const createDevServerConfig = () => {
   return serverConfig;
 };
 
-/*
- * Custom plugin to generate configuration entity yml files for each of our included
- * default Asset Link plugins.
- */
-function GenerateDefaultPluginConfigYmlFilesPlugin() {
-  GenerateDefaultPluginConfigYmlFilesPlugin.prototype.apply = (compiler) => {
-    compiler.hooks.beforeCompile.tap('GenerateDefaultPluginConfigYmlFilesPlugin', (compilation) => {
-
-      const configOutputDir = `${__dirname}/../../../farmos_asset_link/config/install`;
-
-      if (!fs.existsSync(configOutputDir)) {
-        fs.mkdirSync(configOutputDir, { recursive: true });
-      }
-
-      const existingConfigFiles = glob.sync(`${configOutputDir}/farmos_asset_link.asset_link_default_plugin.*.yml`);
-      existingConfigFiles.forEach(f => fs.unlinkSync(f));
-
-      fs.readdirSync(`${__dirname}/plugins`).forEach(filename => {
-        const nameWithoutExt = filename.replace(/(\.[^.]+)*$/, '');
-
-        const configOutputFilename = `${configOutputDir}/farmos_asset_link.asset_link_default_plugin.${nameWithoutExt}.yml`;
-
-        fs.writeFileSync(configOutputFilename, yaml.dump({
-          langcode: 'en',
-          status: true,
-          id: nameWithoutExt,
-          dependencies: { enforced: { module: [ 'farmos_asset_link' ] } },
-          url: `{base_path}alink/plugins/${filename}`,
-          user_defined : null,
-        }));
-      });
-
-    });
-  };
-}
-
 module.exports = {
   output: {
     publicPath: process.env.NODE_ENV === 'production'
@@ -236,7 +201,12 @@ module.exports = {
         },
       ],
     }),
-    new GenerateDefaultPluginConfigYmlFilesPlugin(),
+    new GenerateDefaultPluginConfigYmlFilesPlugin({
+      pluginDir: `${__dirname}/plugins`,
+      configOutputDir: `${__dirname}/../../../farmos_asset_link/config/install`,
+      drupalModuleName: 'farmos_asset_link',
+      pluginUrlFn: (filename) => `{base_path}alink/plugins/${filename}`,
+    }),
   ],
   devServer: createDevServerConfig(),
 };